feat: fixed structure
